requirements.txt
aiohttp==3.7.2
aiohttp-jinja2==1.4.0
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
motor==2.3.0
multidict==5.0.2
pymongo==3.11.0
PyYAML==5.3.1
trafaret==2.1.0
typing-extensions==3.7.4.3
yarl==1.6.3
